+++ lib/builder-interface-augmenter.js
