+++ lib/query/builder.js
@@ -20,7 +20,9 @@ const saveAsyncStack = require('../util/save-async-stack');
 const { isBoolean, isNumber, isObject, isString } = require('../util/is');
 
 const { lockMode, waitMode } = require('./constants');
-const { augmentWithBuilderInterface } = require('../builderInterfaceAugmenter');
+const {
+  augmentWithBuilderInterface,
+} = require('../builder-interface-augmenter');
 
 // Sub-builder for onConflict clauses
 function OnConflictBuilder(builder, columns) {
