+++ lib/raw.js
@@ -11,7 +11,9 @@ const helpers = require('./util/helpers');
 const saveAsyncStack = require('./util/save-async-stack');
 const { nanoid } = require('./util/nanoid');
 const { isNumber, isObject } = require('./util/is');
-const { augmentWithBuilderInterface } = require('./builderInterfaceAugmenter');
+const {
+  augmentWithBuilderInterface,
+} = require('./builder-interface-augmenter');
 
 const debugBindings = debug('knex:bindings');
