+++ lib/schema/builder.js
@@ -3,7 +3,9 @@ const { EventEmitter } = require('events');
 const toArray = require('lodash/toArray');
 const { addQueryContext } = require('../util/helpers');
 const saveAsyncStack = require('../util/save-async-stack');
-const { augmentWithBuilderInterface } = require('../builderInterfaceAugmenter');
+const {
+  augmentWithBuilderInterface,
+} = require('../builder-interface-augmenter');
 
 // Constructor for the builder instance, typically called from
 // `knex.builder`, accepting the current `knex` instance,
